Moved SocketInterface in its own file
